pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "derestrictor"
version = "1.2.0"
description = "Remove refusal behavior from language models using norm-preserving orthogonal projection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["llm", "language-models", "abliteration", "refusal", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "tqdm>=4.65.0",
    "psutil>=5.9.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "click>=8.0.0",
    "huggingface_hub>=0.20.0",
    "safetensors>=0.4.0",
    "datasets>=2.19.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[project.scripts]
derestrictor = "derestrictor.cli.main:cli"
derestrictor-feature-surgery = "derestrictor.cli.feature_surgery:main"
derestrictor-search = "derestrictor.scripts.abliteration_search:main"
derestrictor-quantize = "derestrictor.scripts.batch_quantize:main"
derestrictor-build-dataset = "derestrictor.scripts.build_dataset:main"
derestrictor-upload-dataset = "derestrictor.scripts.upload_dataset:main"
derestrictor-convert-bf16 = "derestrictor.scripts.convert_to_bf16:main"
abliterate = "derestrictor.cli.main:cli"
feature-surgery = "derestrictor.cli.feature_surgery:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.15.11",
    "pre-commit>=4.6.0",
    "sentence-transformers>=2.7.0",
    "pyarrow>=15.0.0",
]

mistral = [
    "mistral-common>=1.8.8",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.15.11",
    "pre-commit>=4.6.0",
    "sentence-transformers>=2.7.0",
    "pyarrow>=15.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/derestrictor"]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [
    "build",
    "dist",
    ".venv",
    ".hf_cache",
    "abliterate",
    "abliterated_models",
    "eval_results",
]

[tool.ruff.lint]
# Professional rule selection. See: https://docs.astral.sh/ruff/rules/
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
    "PERF",   # Perflint
    "PIE",    # flake8-pie
    "RET",    # flake8-return
    "PTH",    # flake8-use-pathlib
    "ARG",    # flake8-unused-arguments
    "ERA",    # eradicate (commented-out code)
    "YTT",    # flake8-2020
    "ASYNC",  # flake8-async
    "FBT",    # flake8-boolean-trap
    "ICN",    # flake8-import-conventions
    "T20",    # flake8-print
    "ISC",    # flake8-implicit-str-concat
    "G",      # flake8-logging-format
    "Q",      # flake8-quotes
    "S",      # flake8-bandit (security)
    "D",      # pydocstyle
]
ignore = [
    "D100",    # Missing module docstring (internal modules don't need them everywhere)
    "D104",    # Missing package docstring in __init__
    "D105",    # Missing docstring in magic method
    "D107",    # Missing docstring in __init__
    "D203",    # Conflicts with D211 (no blank line before class)
    "D213",    # Conflicts with D212 (multi-line summary first line)
    "E501",    # Line length handled by the formatter
    "ISC001",  # Disabled to avoid conflict with the formatter
    "PLR0911", # Too many return statements (decision-heavy ML code)
    "PLR0913", # Too many arguments (common in ML configs)
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison (noisy for numeric ML code)
    "S101",    # Use of assert (fine in research/tests)
    "FBT001",  # Boolean positional arg (common in torch/HF APIs)
    "FBT002",  # Boolean default positional arg
    # ML / research-code idioms
    "G004",    # Logging f-strings: idiomatic and readable; lazy %-formatting is a micro-optimisation here
    "G003",    # Logging string concatenation: same rationale as G004
    "N803",    # Argument names follow linear-algebra conventions (X, W, Q, ...)
    "N806",    # Local variable names follow linear-algebra conventions (W, U, S, V, D, Q, ...)
    "N812",    # Conventional aliases (e.g. `import torch.nn.functional as F`)
    "RUF002",  # Ambiguous Unicode in docstrings: intentional typography (em-dash, non-breaking hyphen)
    "PLC0415", # Imports inside functions: used deliberately for heavy/optional ML deps
]
unfixable = [
    "ERA001",  # Don't auto-delete commented-out code; surface it for review
    "F401",    # Don't auto-delete unused imports; imports may be intentional re-exports
]

[tool.ruff.lint.per-file-ignores]
"src/derestrictor/cli/*.py" = ["T201"]
"src/derestrictor/scripts/*.py" = [
    "T201",    # print() is acceptable in CLI/research utilities
    "S603",    # subprocess call: research scripts are trusted
    "S607",    # start-process with partial path: fine for tooling
]
"src/derestrictor/export/gguf.py" = [
    "S603",    # subprocess invokes vetted llama.cpp binaries with constructed args
]
# eval/ modules double as CLI tools (they have their own argparse main()),
# so print() is acceptable and subprocess-adjacent lints are relaxed.
"src/derestrictor/eval/*.py" = [
    "T201",    # print() is acceptable in CLI/research utilities
    "S603",    # subprocess call: research scripts are trusted
    "S607",    # start-process with partial path: fine for tooling
]
"src/derestrictor/eval/harness.py" = ["S", "D", "PLR", "ARG", "T201"]
"tests/**/*.py" = ["S", "D", "PLR", "ARG"]
"**/__init__.py" = ["F401", "D"]

[tool.ruff.lint.isort]
known-first-party = ["derestrictor"]
combine-as-imports = true
split-on-trailing-comma = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
markers = [
    "requires_cuda: test requires a CUDA device and is skipped on CPU-only runners",
]
